refactored graph.js to seperate page from graph
