+++ client/package.json
@@ -9,12 +9,13 @@
     "@mui/material": "^5.10.13",
     "bootstrap": "^5.2.2",
     "cra-template": "1.2.0",
+    "d3": "^7.6.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-force-graph": "^1.41.14",
-    "react-scripts": "^2.1.3",
     "react-router-dom": "^6.4.3",
+    "react-scripts": "^2.1.3",
     "react-slider": "^2.0.4"
   },
   "scripts": {
