--- client/src/Graph.js
@@ -1,139 +1,0 @@
-import React, { useEffect, useRef, useState, useCallback } from "react";
-import "./App.css";
-import "bootstrap/dist/css/bootstrap.min.css";
-import { ForceGraph2D } from "react-force-graph";
-import TimelineNavigation from "./Slider";
-import { useParams } from "react-router-dom";
-
-// Converts JSON data from backend into graph JSON data for react force graph
-function convert(data) {
-  let result = [...data["sections"]];
-  return result.map(convertToGraph);
-}
-
-function convertToGraph(data) {
-  let nodes = [];
-  let links = [];
-  let scale = 10;
-  let names = data["names"];
-  let matrix = data["matrix"];
-  for (let i = 0; i < names.length; i++) {
-    nodes.push({ id: "id" + String(i), name: names[i] });
-  }
-  for (let j = 0; j < names.length; j++) {
-    for (let k = 0; k < names.length; k++) {
-      if (j !== k && matrix[j][k] !== 0) {
-        links.push({
-          source: "id" + String(j),
-          target: "id" + String(k),
-          value: matrix[j][k] * scale,
-          linkVisibility: true,
-        });
-      }
-    }
-  }
-  return { nodes: nodes, links: links };
-}
-
-function Graph() {
-  const [displayWidth, setDisplayWidth] = useState(window.innerWidth);
-  const [displayHeight, setDisplayHeight] = useState(window.innerHeight);
-
-  window.addEventListener("resize", () => {
-    setDisplayWidth(window.innerWidth);
-    setDisplayHeight(window.innerHeight);
-  });
-
-  const [data, setData] = useState({ nodes: {}, links: {} });
-  const [counter, setCounter] = useState(0);
-  const [activeNode, setActiveNode] = useState("");
-
-  // Fetches data outputted by the backend
-  const params = useParams();
-
-  const getData = () => {
-    fetch(params.book + ".json", {
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-    })
-      .then(function (response) {
-        return response.json();
-      })
-      .then(function (myJson) {
-        let res = convert(myJson);
-        setData(res);
-      });
-  };
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  // Rendering Graph
-
-  const forceRef = useRef(null);
-  let repelStrength = -10;
-  let centering = -120;
-  let zoomingTime = 50;
-  let padding = 130;
-  let widthCentering = 300;
-  let heightCentering = 100;
-
-  useEffect(() => {
-    forceRef.current.d3Force("charge").strength(repelStrength);
-    forceRef.current.d3Force("center").x(centering);
-    forceRef.current.zoomToFit(zoomingTime, padding);
-    document.body.style.backgroundColor = "#eae0d5";
-  });
-
-  console.log(data);
-
-  const handleClick = useCallback((node) => {
-    if (activeNode === "") {
-      setActiveNode(node["id"]);
-    } else if (activeNode === node["id"]) {
-      setActiveNode("");
-    } else {
-      return;
-    }
-
-    let links = data[counter]["links"];
-    console.log(links);
-    for (let i = 0; i < links.length; i++) {
-      if (links[i]["source"]["id"] !== node["id"]) {
-        links[i]["linkVisibility"] = !links[i]["linkVisibility"];
-      }
-    }
-  });
-
-  return (
-    <>
-      <div className="App">
-        <ForceGraph2D
-          graphData={data[counter]}
-          nodeLabel="name"
-          linkCurvature="curvature"
-          linkWidth="value"
-          linkVisibility="linkVisibility"
-          linkDirectionalParticleWidth={1}
-          width={displayWidth - widthCentering}
-          height={displayHeight - heightCentering}
-          ref={forceRef}
-          nodeAutoColorBy={"name"}
-          onNodeClick={handleClick}
-        />
-
-        <TimelineNavigation
-          maxval={data.length ? data.length - 1 : 0}
-          setCounter={setCounter}
-          counter={counter}
-          TimelineNavigation
-        />
-      </div>
-    </>
-  );
-}
-
-export default Graph;
